Refactor: 중복 존재했던 enum 제거 및 TVSERIES로 이름 변경
- enum 폴더를 enums로 변경했습니다.

- types 폴더에도 똑같은 enum이 존재해서 제거했습니다. (enums 폴더에 존재하는 enum 사용)

- ContentType의 TV를 TVSERIES로 바꿨습니다.
