+++ src/common/enums/content-type.enum.ts
@@ -1,4 +1,4 @@
 export enum ContentType {
   MOVIE = 'movie',
-  TV = 'tv',
+  TVSERIES = 'tvSeries',
 }
