--- src/common/types/content-type.enum.ts
@@ -1,4 +1,0 @@
-export enum ContentType {
-  MOVIE = 'movie',
-  TV = 'tv',
-}
