--- src/common/types/source-type.enum.ts
@@ -1,4 +1,0 @@
-export enum SourceType {
-  TMDB = 'tmdb',
-  YOUTUBE = 'youtube',
-}
