+++ src/content/movie/dto/movie-list-response.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 import { Expose, Type } from 'class-transformer';
-import { ContentType } from 'src/common/types/content-type.enum';
+import { ContentType } from 'src/common/enums/content-type.enum';
 import { MovieListItemDto } from './movie-list-item.dto';
 
 export class MovieListResponseDto {
